journald: never block when sending messages on NOTIFY_SOCKET socket
Otherwise we might run into deadlocks, when journald blocks on the
notify socket on PID 1, and PID 1 blocks on IPC to dbus-daemon and
dbus-daemon blocks on logging to journald. Break this cycle by making
sure that journald never ever blocks on PID 1.

Note that this change disables support for event loop watchdog support,
as these messages are sent in blocking style by sd-event. That should
not be a big loss though, as people reported frequent problems with the
watchdog hitting journald on excessively slow IO.

Fixes: #1505.

Resolves: #1425758
Cherry-picked from: e22aa3d3284709234f086ebebc13a905a295b7a7

[msekleta: note that this version of the fix doesn't contain code that
sends stream fds via notify socket. This is because our journald version
doesn't have this capability. It was reverted because it caused weird
issues on upgrade from v208 to v219 that we did in RHEL-7.2]